requirements.txt
# TODO:
-i https://pypi.org/simple
django==2.2.6
djangorestframework==3.10.4